the appraisal variables of desirability, desirability for other, and praiseworthiness used in emotional reaction rules were changed from integers to floats
git-svn-id: http://svn.lirec.eu@284 dcc22606-d844-4505-a083-a3dbd7937ca2

+++ AgentMind/trunk/AgentMind/FAtiMA/src/FAtiMA/util/parsers/CultureLoaderHandler.java
@@ -133,10 +133,10 @@ public class CultureLoaderHandler extends ReflectXMLHandler {
 		Event event = Event.ParseEvent(subject, action, target, parameters);
 
 		//Create the reaction
-		Integer desirability = new Integer(0);
-		Integer desirabilityForOther = new Integer(0);
+		Float desirability = new Float(0);
+		Float desirabilityForOther = new Float(0);
 		Symbol other = null;
-		Integer praiseworthiness = new Integer(attributes.getValue("value"));
+		Float praiseworthiness = new Float(attributes.getValue("value"));
 		Reaction _eventReaction = new Reaction(desirability, desirabilityForOther, praiseworthiness, other);
 		_eventReaction.setEvent(event);
